# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: صرح - Career Development Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with national ID and Nafath
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Not authenticated
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/cv/analyze:
    post:
      operationId: analyzeCv
      tags: [profile]
      summary: Trigger CV analysis after upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CvAnalyzeInput"
      responses:
        "200":
          description: CV analyzed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvAnalysis"
...
  /career/certifications:
    get:
      operationId: getCertifications
      tags: [career]
      summary: Get linked academic and professional certifications
      responses:
        "200":
          description: Certifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certification"
    post:
      operationId: addCertification
      tags: [career]
      summary: Add a certification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationInput"
      responses:
        "201":
          description: Certification added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
  /career/activities:
    get:
      operationId: getActivities
      tags: [career]
      summary: Get extracurricular activities (clubs, external centers)
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: addActivity
      tags: [career]
      summary: Add an activity with proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Activity added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send message to AI career advisor chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [nationalId, fullName]
      properties:
        nationalId:
          type: string
          description: Saudi national ID (السجل المدني)
        fullName:
          type: string
          description: Full name in Arabic
        nafathVerified:
          type: boolean
          description: Whether Nafath verification was completed
          default: false
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserSession"
        token:
          type: string
        isNewUser:
          type: boolean
    UserSession:
      type: object
      required: [id, nationalId, fullName]
      properties:
        id:
          type: integer
        nationalId:
          type: string
        fullName:
          type: string
        profileComplete:
          type: boolean
        avatarInitials:
          type: string
    Profile:
      type: object
      required: [id, nationalId, fullName]
      properties:
        id:
          type: integer
        nationalId:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        careerGoal:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        major:
          type: ["string", "null"]
        gpa:
          type: ["number", "null"]
        interests:
          type: array
          items:
            type: string
        hobbies:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
        cvFileName:
          type: ["string", "null"]
        cvAnalyzed:
          type: boolean
        desiredSkills:
          type: array
          items:
            type: string
        profileCompleteness:
          type: integer
        createdAt:
          type: string
    ProfileInput:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        careerGoal:
          type: string
        currentRole:
          type: string
        educationLevel:
          type: string
        university:
          type: string
        major:
          type: string
        gpa:
          type: number
        interests:
          type: array
          items:
            type: string
        hobbies:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
        desiredSkills:
          type: array
          items:
            type: string
    Language:
      type: object
      required: [name, level]
      properties:
        name:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, native]
    CvAnalyzeInput:
      type: object
      required: [fileName]
      properties:
        fileName:
          type: string
        fileContent:
          type: string
          description: Base64-encoded file content
    CvAnalysis:
      type: object
      required: [fileName, extractedSkills, summary]
      properties:
        fileName:
          type: string
        extractedSkills:
          type: array
          items:
            type: string
        summary:
          type: string
        experienceYears:
          type: ["number", "null"]
        educationFound:
          type: ["string", "null"]
...
    Certification:
      type: object
      required: [id, title, issuer, type]
      properties:
        id:
          type: integer
        title:
          type: string
        issuer:
          type: string
        type:
          type: string
          enum: [academic, professional, language, government]
        score:
          type: ["string", "null"]
        dateEarned:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        verificationProof:
          type: ["string", "null"]
        verified:
          type: boolean
        platform:
          type: ["string", "null"]
    CertificationInput:
      type: object
      required: [title, issuer, type]
      properties:
        title:
          type: string
        issuer:
          type: string
        type:
          type: string
          enum: [academic, professional, language, government]
        score:
          type: string
        dateEarned:
          type: string
        expiryDate:
          type: string
        verificationProof:
          type: string
        platform:
          type: string
    Activity:
      type: object
      required: [id, title, organization, type]
      properties:
        id:
          type: integer
        title:
          type: string
        organization:
          type: string
        type:
          type: string
          enum: [university_club, external_center, volunteer, competition, internship]
        role:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        proofFile:
          type: ["string", "null"]
        verified:
          type: boolean
    ActivityInput:
      type: object
      required: [title, organization, type]
      properties:
        title:
          type: string
        organization:
          type: string
        type:
          type: string
          enum: [university_club, external_center, volunteer, competition, internship]
        role:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: string
        proofFile:
          type: string
    ChatMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        context:
          type: string
    ChatResponse:
      type: object
      required: [id, reply, timestamp]
      properties:
        id:
          type: integer
        reply:
          type: string
        timestamp:
          type: string
        suggestions:
          type: array
          items:
            type: string
...