# lib/api-zod/src/generated/types/cvAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * صرح - Career Development Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CvAnalysis {
  fileName: string;
  extractedSkills: string[];
  summary: string;
  /** @nullable */
  experienceYears?: number | null;
  /** @nullable */
  educationFound?: string | null;
}